ci: Add Python test workflow
This commit adds a GitHub Actions CI workflow that runs the full
testsuite with the "release test" script using various Python versions
and host operating systems.

Note that the testing on Windows host is currently disabled because the
test scripts do not correctly handle Windows paths at this time.

Also note that Python 2.7 is not tested because it is now archaic and
supporting it is pointless.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>