guiconfig: Add show all checkbox in jump to dialog
Invisible items in jump to make it hard to find a configurable item if
the keyword is widely used, like LOG_LEVEL_XXX.
Add a show all check box to allow user toggle invisiable result in jump
to matches. It was disabled by default, aligned with the show all button
in main menu.

Signed-off-by: Yves Wang <zhengjia.wang@nxp.com>